bate mnogo buckam (beshe ChatGPT)

+++ install.sh
@@ -6,6 +6,11 @@ if [ ! -d $PLUGIN_DIR ]; then
 	git clone https://github.com/tmux-plugins/tpm $PLUGIN_DIR
 fi
 
-ln -s $HOME/tmux-conf/.tmux.conf $HOME/.tmux.conf
+TMUX_DEST=$HOME/.tmux.conf
+
+if ! test -f $TMUX_DEST; then
+	ln -s $HOME/tmux-conf/.tmux.conf $HOME/.tmux.conf
+fi
+
 
 tmux source-file $HOME/.tmux.conf
